+++ cloud/src/services/index.js
@@ -2,16 +2,16 @@ const Batch = require('./batch/batch.js');
 const Roles = require('./roles/roles.js');
 const Aggregate = require('./aggregate/aggregate.js');
 const Post = require('./post/post.js');
-const Messaging = require('./messaging/messaging.js');
 const Offline = require('./offline/offline.js');
+const Messaging = require('./messaging/messaging');
 
-const services = {};
-
-services.batch = Batch;
-services.roles = Roles;
-services.aggregate = Aggregate;
-services.post = Post;
-services.messaging = Messaging;
-services.offline = Offline;
+const services = {
+  batch: Batch,
+  roles: Roles,
+  aggregate: Aggregate,
+  post: Post,
+  Messaging,
+  offline: Offline,
+};
 
 module.exports = services;